there can be but one
